Types fix: first pass
 This pass makes variable size_t-ready, so everything works when
the prototypes are fixed in skalibs.
 Some code uses "unsigned int *" where it should be "size_t *";
it cannot be changed now, but it's been marked with XXX. It must
change at the same time as the skalibs API.

+++ src/execline/backtick.c
@@ -19,7 +19,7 @@ int main (int argc, char const **argv, char const *const *envp)
   pid_t pid ;
   int argc1, fdwstat ;
   stralloc modif = STRALLOC_ZERO ;
-  unsigned int modifstart ;
+  size_t modifstart ;
   int insist = 0, chomp = 0 ;
   char const *def = 0 ;
   PROG = "backtick" ;
@@ -74,7 +74,7 @@ int main (int argc, char const **argv, char const *const *envp)
   if (argc1 == argc - 1) return 0 ;
   if (!stralloc_0(&modif)) strerr_diefu1sys(111, "stralloc_catb") ;
   {
-    unsigned int reallen = str_len(modif.s) ;
+    size_t reallen = str_len(modif.s) ;
     if (reallen < modif.len - 1)
     {
       if (insist >= 2)
